+++ libs/plugin-retry/src/lib/predicates/is-server-error.ts
@@ -1,3 +1,3 @@
-import { RetryPredicate } from '@http-ext/plugin-retry';
+import { RetryPredicate } from './retry-predicate';
 
 export const isServerError: RetryPredicate = response => response.status >= 500;
