+++ libs/plugin-retry/src/lib/predicates/is-server-or-unknown-error.ts
@@ -1,0 +1,13 @@
+import { RetryPredicate } from './retry-predicate';
+import { isServerError } from './is-server-error';
+import { isUnknownError } from './is-unknown-error';
+
+export const or = <TArgs extends any[]>(
+  ...predicates: ((...args: TArgs) => boolean)[]
+) => (...args: TArgs) => {
+  return predicates.some(predicate => predicate(...args));
+};
+
+export const isServerOrUnknownError: RetryPredicate = response => {
+  return or(isServerError, isUnknownError)(response);
+};
