+++ libs/plugin-retry/src/lib/predicates/is-unknown-error.ts
@@ -1,3 +1,3 @@
-import { RetryPredicate } from '@http-ext/plugin-retry';
+import { RetryPredicate } from './retry-predicate';
 
 export const isUnknownError: RetryPredicate = response => response.status === 0;
